Sources synchronized with the dovelopment of the build version 3.1.004.007

+++ src/test/java/test/xdef/component/N_i.java
@@ -4,8 +4,8 @@
 package test.xdef.component;
 
 public interface N_i extends cz.syntea.xdef.component.XComponent {
-  public Long getOne();
-  public void setOne(Long x);
+  public Integer getOne();
+  public void setOne(Integer x);
   public String xposOfOne();
   public String getTwo();
   public void setTwo(String x);
